added aladded all common elements from auth confirmations to CommonConfirmationPage
