+++ src/common/CommonConfirmationPage/CommonConfirmationPage.module.css
@@ -29,6 +29,31 @@
   display: none;
 }
 
+.confirmation-content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+}
+
+.confirmation-icon {
+  width: 125px !important;
+  height: 125px !important;
+  padding: 5px;
+  border: 7.5px solid #345e80;
+  border-radius: 50%;
+}
+
+.confirmation-message {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  margin: 50px;
+}
+
 @media screen and (max-width: 768px) {
   .logo {
     display: none;
@@ -36,5 +61,15 @@
 
   .logo-mobile {
     display: block;
+  }
+
+  .confirmation-title {
+    font-size: 28px !important;
+  }
+
+  .confirmation-button {
+    position: absolute !important;
+    bottom: 60px;
+    width: 80% !important;
   }
 }
