--- src/components/ForgotPassword/ForgotPasswordConfirmation/ForgotPasswordConfirmation.module.css
@@ -1,36 +1,0 @@
-.forgot-password-confirmation-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-.confirmation-icon {
-  width: 125px !important;
-  height: 125px !important;
-  padding: 5px;
-  border: 7.5px solid #345e80;
-  border-radius: 50%;
-}
-
-.confirmation-message {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  margin: 50px;
-}
-
-@media screen and (max-width: 768px) {
-  .confirmation-header {
-    font-size: 28px !important;
-  }
-
-  .confirmation-button {
-    position: absolute !important;
-    bottom: 60px;
-    width: 80% !important;
-  }
-}
